tests/wpt/web-platform-tests/html/browsers/the-window-object/the-windowproxy-object/test-window-proxy-statusbar-manual.html

<!doctype html>
<html>
<head>
    <meta charset="utf-8" />
    <title>WindowProxy statusbar visible flag Test</title>
    <link rel="author" title="dokenzy" href="dokenzy@gmail.com">
    <script type="text/javascript" src="/resources/testharness.js"></script>
    <script type="text/javascript" src="/resources/testharnessreport.js"></script>
</head>
<body>
    <h1>Description</h1>
    <p>WindowProxy.statusbar Test</p>

    <h1>Manual Test Steps:</h1>
    <ol>
        <li>Make the statusbar visible in the user agent before executing this test.</li>
        <li>You may need to manually reload afterwards.</li>
    </ol>

    <div id="log"></div>

    <script type="text/javascript" >
    test(function () {
        assert_not_equals(typeof window.statusbar.visible, undefined, 'window.statusbar.visible');
        assert_true(window.statusbar.visible, 'window.statusbar.visible');
    }, "BarProp attribute: window.statusbar.visible");
    </script>
</body>
</html>
